requirements-base.txt
# Base requirements for Django LangChain Story Generator

# Django and DRF
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
redis==5.0.4

# LangChain and AI
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
transformers==4.42.3
torch==2.3.1
huggingface-hub==0.23.4

# Image Processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# Audio Processing
SpeechRecognition==3.10.4
pydub==0.25.1
pyaudio==0.2.14

# HTTP Client
requests==2.32.3
httpx==0.27.0

# File Processing
python-magic==0.4.27

# Validation
pydantic==2.8.2

# Environment
python-decouple==3.8
python-dotenv==1.0.1

# Monitoring and Logging
celery==5.3.7
flower==2.0.1

# Testing
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0
